refactor: Post validation 로직 수정 및 메서드명 리팩토링
