+++ src/test/java/com/backendoori/ootw/post/domain/PostTest.java
@@ -23,21 +23,23 @@ class PostTest {
     private static final String IMG_URL = "imgUrl";
     private static final User MOCK_USER = mock(User.class);
     private static final TemperatureArrangeDto TEMPERATURE_ARRANGE_DTO = new TemperatureArrangeDto(-10.0, 10.0);
-    static final TemperatureArrange TEMPERATURE_ARRANGE = TemperatureArrange.from(TEMPERATURE_ARRANGE_DTO);
+    private static final TemperatureArrange TEMPERATURE_ARRANGE = TemperatureArrange.from(TEMPERATURE_ARRANGE_DTO);
 
     private static Stream<Arguments> provideInvalidInfo() {
         return Stream.of(Arguments.of("title이 null인 경우",
-                new PostSaveRequest(null, "Test Content", NX, NY)),
+                new PostSaveRequest(null, "Test Content", NX, NY), TEMPERATURE_ARRANGE),
             Arguments.of("title이 공백인 경우",
-                new PostSaveRequest(" ", "Test Content", NX, NY)),
+                new PostSaveRequest(" ", "Test Content", NX, NY), TEMPERATURE_ARRANGE),
             Arguments.of("title이 30자를 넘는 경우",
-                new PostSaveRequest("T".repeat(31), "Test Content", NX, NY)),
+                new PostSaveRequest("T".repeat(31), "Test Content", NX, NY), TEMPERATURE_ARRANGE),
             Arguments.of("content가 null인 경우",
-                new PostSaveRequest("Test Title", null, NX, NY)),
+                new PostSaveRequest("Test Title", null, NX, NY), TEMPERATURE_ARRANGE),
             Arguments.of("content가 공백인 경우",
-                new PostSaveRequest("Test Title", " ", NX, NY)),
+                new PostSaveRequest("Test Title", " ", NX, NY), TEMPERATURE_ARRANGE),
             Arguments.of("content가 500자를 넘는 경우",
-                new PostSaveRequest("Test Title", "T".repeat(501), NX, NY))
+                new PostSaveRequest("Test Title", "T".repeat(501), NX, NY), TEMPERATURE_ARRANGE),
+            Arguments.of("temperatureArrange가 null인 경우",
+                new PostSaveRequest("Test Title", "T".repeat(501), NX, NY), null)
         );
     }
 
@@ -62,10 +64,10 @@ class PostTest {
     @ParameterizedTest(name = "[{index}] {0}")
     @MethodSource("provideInvalidInfo")
     @DisplayName("from 메서드로 유효하지 않은 User, PostSaveRequest로부터 Post를 생성하는 것에 실패한다.")
-    void createPostFail(String info, PostSaveRequest postSaveRequest) {
+    void createPostFail(String info, PostSaveRequest postSaveRequest, TemperatureArrange temperatureArrange) {
         // given // when, then
         assertThrows(IllegalArgumentException.class,
-            () -> Post.from(MOCK_USER, postSaveRequest, IMG_URL, TEMPERATURE_ARRANGE));
+            () -> Post.from(MOCK_USER, postSaveRequest, IMG_URL, temperatureArrange));
     }
 
 }
